refactor(codex): isolate MCP CLI argument serialization
