+++ packages/web/src/components/MoveCardModal.vue
@@ -92,7 +92,6 @@
             </label>
           </div>
         </div>
-
       </div>
 
       <div class="modal-footer">
